Fix crashing for unknown notification id. (#36)
A package can be tracked (non-null list value from map) but still have posted a notification before the service started listening to (tracking) notification ids. So, we might get a onNotificationRemoved call with a package we are tracking but a notification id we don't know about.
That's a valid state.